Updating buttons to be cleaner. Starting support for nested buttons and inheritance. Plans to support cascading style settings.
