+++ entities/x/btn/textBox.go
@@ -23,6 +23,10 @@ func (b *TextBox) Init() event.CID {
 func NewTextBox(cid event.CID, x, y, w, h, txtX, txtY float64,
 	f *render.Font, r render.Renderable, layers ...int) *TextBox {
 
+	if f == nil {
+		f = render.DefFont()
+	}
+
 	b := new(TextBox)
 
 	cid = cid.Parse(b)
